Fixed issue with constructor of ElementArrayFinderWrapper and circular dependency between ElementFinderWrapper and ElementArrayFinderWrapper
