+++ src/Eyes.js
@@ -20,14 +20,13 @@
         EyesWebDriver = require('./EyesWebDriver'),
         EyesRemoteWebElement = require('./EyesRemoteWebElement'),
         EyesWebDriverScreenshot = require('./EyesWebDriverScreenshot'),
-        ElementFinderWrappers = require('./ElementFinderWrapper'),
+        ElementFinderWrapper = require('./ElementFinderWrappers').ElementFinderWrapper,
+        ElementArrayFinderWrapper = require('./ElementFinderWrappers').ElementArrayFinderWrapper,
         ScrollPositionProvider = require('./ScrollPositionProvider'),
         CssTranslatePositionProvider = require('./CssTranslatePositionProvider'),
         ElementPositionProvider = require('./ElementPositionProvider'),
         EyesRegionProvider = require('./EyesRegionProvider');
-    var ElementFinderWrapper = ElementFinderWrappers.ElementFinderWrapper,
-        ElementArrayFinderWrapper = ElementFinderWrappers.ElementArrayFinderWrapper,
-        EyesBase = EyesSDK.EyesBase,
+    var EyesBase = EyesSDK.EyesBase,
         FixedScaleProvider = EyesSDK.FixedScaleProvider,
         ContextBasedScaleProviderFactory = EyesSDK.ContextBasedScaleProviderFactory,
         FixedScaleProviderFactory = EyesSDK.FixedScaleProviderFactory,
